+++ index.html
@@ -11,7 +11,11 @@
     <div class="container">
       <header class="header">
         <div class="img-container">
-          <img src="/" alt="Rick and Morty logo" class="header__img" />
+          <img
+            src="./src/logo.jpeg"
+            alt="Rick and Morty logo"
+            class="header__img"
+          />
         </div>
         <div class="search">
           <select class="header__filter" name="status" id="status">
