+++ src/main.js
@@ -1,5 +1,5 @@
-import { createCharacter } from "./lib/character.js";
-import { logoSetup } from "./lib/logo.js";
+import { createCharacter } from "./lib/api.js";
+import { clearCharacters } from "./lib/api.js";
 
 const fetchButton = document.querySelector(".header__search-button");
 
@@ -7,10 +7,3 @@ fetchButton.addEventListener("click", () => {
   clearCharacters();
   createCharacter();
 });
-
-function clearCharacters() {
-  const characterElements = document.querySelectorAll(".character");
-  characterElements.forEach((section) => section.remove());
-}
-
-logoSetup();
