+++ src/styles/_base.css
@@ -31,6 +31,7 @@ html {
 
 .header__img {
   width: 60px;
+  border-radius: 50%;
 }
 
 .search {
